adding /map2 visual from planner

+++ Simulation/catkin_ws/src/global_planner/sbpl_lattice_planner/sbpl/src/include/sbpl/planners/araplanner.h
@@ -289,7 +289,7 @@ protected:
     ARASearchStateSpace_t* pSearchStateSpace_;
 
     unsigned int searchexpands;
-    int MaxMemoryCounter;
+    long long int MaxMemoryCounter;
     clock_t TimeStarted;
     FILE *fDeb;
